doc added

+++ todo-backend/src/main/java/org/min/todo/controller/UserController.java
@@ -27,27 +27,61 @@ public class UserController {
         return ResponseEntity.ok(userService.login(dto));
     }
 
+    /**
+     * @Method Name : updateUser
+     * @작성일 : 2021. 08. 30.
+     * @작성자 : J.M YOON
+     * @변경이력 :
+     * @Method 설명 : Security Context Holder 안에 있는 유저정보와 dto의 유저 정보가 같으면 유저 정보를 수정한다.
+     * @param dto 변경하는 대상과 내용
+     * @param principalDetails 요청한 대상
+     * @return 변경된 유저 정보
+     */
     @PutMapping("/update")
     public ResponseEntity<UserDto> updateUser(@RequestBody UserDto dto,@AuthenticationPrincipal PrincipalDetails principalDetails) {
-        if(principalDetails.getUsername().equals(dto.getUsername())) {
-            return ResponseEntity.ok(userService.modify(dto));
-        }
+        if(principalDetails.getUsername().equals(dto.getUsername())) return ResponseEntity.ok(userService.modify(dto));
         throw new IllegalArgumentException("권한이 없습니다.");
     }
 
+    /**
+     * @Method Name : getUser
+     * @작성일 : 2021. 08. 30.
+     * @작성자 : J.M YOON
+     * @변경이력 :
+     * @Method 설명 : 선택한 유저의 정보를 가져온다.
+     * @param username 선택할 유저의 아이디
+     * @return 가져온 유저 정보
+     */
     @GetMapping("/get/user/{username}")
     public ResponseEntity<UserDto> getUser(@PathVariable String username) {
         return ResponseEntity.ok(userService.getUser(username));
     }
 
+    /**
+     * @Method Name : updateUser
+     * @작성일 : 2021. 08. 30.
+     * @작성자 : J.M YOON
+     * @변경이력 :
+     * @Method 설명 : Security Context Holder 안에 있는 유저정보와 삭제하려는 유저의 정보가 같으면 유저 정보를 삭제한다.
+     * @param username 삭제하려는 유저의 아이디
+     * @param principalDetails 요청한 대상
+     * @return 삭제여부
+     */
     @DeleteMapping("/delete/{username}")
     public ResponseEntity<String> removeUser(@PathVariable String username, @AuthenticationPrincipal PrincipalDetails principalDetails) {
-        if(principalDetails.getUsername().equals(username)) {
-            return ResponseEntity.ok(userService.remove(username));
-        }
+        if(principalDetails.getUsername().equals(username)) return ResponseEntity.ok(userService.remove(username));
         throw new IllegalArgumentException("권한이 없습니다.");
     }
 
+    /**
+     * @Method Name : refreshToken
+     * @작성일 : 2021. 08. 30.
+     * @작성자 : J.M YOON
+     * @변경이력 :
+     * @Method 설명 : accessToken을 새로 발급해준다.
+     * @param dto accessToken 정보와 유저 아이디
+     * @return 새로 발급한 accessToken과 유저 아이디
+     */
     @PostMapping("/refresh")
     public ResponseEntity<TokenDto> refreshToken(@RequestBody TokenDto dto) {
         return ResponseEntity.ok(userService.refreshToken(dto));
